front_end/requirements.txt
pandas
streamlit==1.14.0
streamlit-option-menu==0.3.2
Pillow==9.0.1
